+++ lib/bundler/source.rb
@@ -15,6 +15,12 @@ module Bundler
       mirrors[normalized_key] || uri
     end
 
+    attr_accessor :dependency_names
+
+    def unmet_deps
+      specs.unmet_dependency_names
+    end
+
     def version_message(spec)
       locked_spec = Bundler.locked_gems.specs.find { |s| s.name == spec.name } if Bundler.locked_gems
       locked_spec_version = locked_spec.version if locked_spec
@@ -25,5 +31,8 @@ module Bundler
       message
     end
 
+    def can_lock?(spec)
+      spec.source == self
+    end
   end
 end